CCF branded color palettes

Creates color palettes based on CCF brand colors, including "main", which contains all primary and secondary CCF colors, "bright", which contains all supporting bright CCF colors, "neutral", which contains all supporting neutral CCF colors, "all", which contains all colors from main, bright, and neutral for when more colors are needed, "blues", which contains a high and low blue color to create continuous blue scales, "greens", which contains a high and low green color to create continuous green scales

ccf_palettes

Format

An object of class list of length 7.

Examples

ccf_palettes[["main"]]
#> [1] "#0078bf" "#00843d" "#4b4b45" "#f08122" "#5c2161" "#a61f56" "#f7c612"
ccf_palettes[["contrast"]]
#>  [1] "#0078bf" "#f7c612" "#a61f56" "#00843d" "#f08122" "#5c2161" "#d5e048"
#>  [8] "#ef413d" "#79d0e6" "#7b7d83"